Fix the `states.iso2` and `states.type` values for Jämtland County in Sweden
`states.iso2` for Jämtland County in Sweden (ID 5117) is `0` and `type` is
`SE-Z`, which are probably mistakes.

This changes `states.iso2` to `Z`, which is the county code for Jämtland (see
https://en.wikipedia.org/wiki/Counties_of_Sweden). The `SE-` prefix isn't used
since no other Swedish counties use it (see ID 1542, for example).

This also changes `states.type` to `NULL`, since no other Swedish counties
specify a type.